# lib/api-zod/src/generated/types/adPostsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Meisterverbund API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdPost } from "./adPost";

export interface AdPostsListResponse {
  posts: AdPost[];
  total: number;
  page: number;
  totalPages: number;
}